requirements.txt
amqp==5.3.1
asgiref==3.8.1
astroid==3.2.3
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
autopep8==2.3.1
azure-core==1.32.0
azure-storage-blob==12.24.0
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==43.0.3
daphne==4.1.2
dill==0.3.8
Django==5.0.6
django-cors-headers==4.4.0
django-environ==0.11.2
django-phonenumber-field==8.0.0
django-storages==1.14.4
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-extra-fields==3.7.0
filetype==1.2.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
isodate==0.7.2
isort==5.13.2
kombu==5.4.2
mccabe==0.7.0
msgpack==1.1.0
phonenumbers==8.13.50
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.48
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.0
pycparser==2.22
PyJWT==2.8.0
pylint==3.2.5
pyOpenSSL==24.2.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
razorpay==1.4.2
redis==5.2.1
requests==2.32.3
service-identity==24.2.0
setuptools==75.1.0
six==1.16.0
sqlparse==0.5.0
tomlkit==0.13.0
Twisted==24.10.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
zope.interface==7.1.1
